pyproject.toml
[project]
name = "v"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docstring-parser>=0.16",
    "loguru>=0.7.3",
    "mcp[cli]>=1.6.0",
    "openai>=1.75.0",
    "pillow>=11.2.1",
    "polars>=1.27.1",
    "pydantic>=2.11.3,<3",
    "ruff>=0.11.6",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
]
